Refactoring ZKClient to use interface + builder pattern and hide the implementation details.
